rename FindNewestFile.py to utils.py in preparation for other utility functions
