+++ rtAttenPy/utils.py
